% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocWrapperTM.R
\name{prepPredictionLDA}
\alias{prepPredictionLDA}
\title{preprocessor on prediction}
\usage{
prepPredictionLDA(data, target, args, control = list(k, exclude, model))
}
\description{
preprocessor on prediction
}
