% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocWrapperTM.R
\name{prepTrainLDA}
\alias{prepTrainLDA}
\title{topic model (LDA) wrapper for mlr preprocessing
preprocessor on training}
\usage{
prepTrainLDA(data, target, args = list(k, exclude))
}
\description{
topic model (LDA) wrapper for mlr preprocessing
preprocessor on training
}
